+++ INF4710 - Introduction Multimedia/TP2/inf4710_tp2_cpp/dct.h
@@ -14,7 +14,7 @@ inline int uv_2_index(int uv) {
 	return uv - 1;
 }
 
-inline float getC(int w, int n)
+inline double getC(int w, int n)
 {
 	if (w == 1)
 	{
